Handle optionalDependencies.
When a package is installed by `npm`, any of its `optionalDependencies`
are added to the packages own `dependencies` hash. Unfortunately, this
means that we will discover things in `dependencies` that are not
_actually_ present in the `node_modules` heirarchy.

This commit handles the specific error condition when a package cannot
be found and simply returns `null`, if the package is missing it (by
definition) cannot affect the host packages cache key (which is why it
is ok to just ignore missing packages).