Fix issues with resolve.sync when using absolute paths.
Prior to these changes, we returned the passed in `name` param directly
when it started with a `/`. This meant that _sometimes_ `resolvePkg`
returned a directory path (when it was provided an absolute path) and
other times it would return a path to the `package.json` file. We would
subsequently strip off the trailing 12 chars (to remove `package.json`)
for the path to `statPathsFor` resulting in completely invalid paths to
stat when inputs were an absolute path.