[BUGFIX] restore old behavior – hashing garbage input should not error
hashForDep with invalid input, would not error rather would produce a checksum. So restoring this for now, although we should remove in the next major version.

One deviation from the old behavior is to ensure the all garbage input doesn’t produce the same hash, rather it will get a hash based on inputs provided.

Context:

* https://github.com/emberjs/ember.js/pull/17679
* https://github.com/emberjs/ember.js/issues/17678